requirements.txt
Flask
pymongo
dnspython
google-genai
python-dotenv